Pi Pico lora instructions
Raspberry Pi Pico board
LoRa transceiver module (e.g., SX1278 or SX1276)
Breadboard
Jumper wires
USB cable
Computer running a terminal program
Here are the steps to send and receive messages using LoRa and Raspberry Pi Pico:

Connect the LoRa transceiver module to the Raspberry Pi Pico board as follows:

VCC of LoRa module to 3.3V of Raspberry Pi Pico
GND of LoRa module to GND of Raspberry Pi Pico
MISO of LoRa module to GP16 of Raspberry Pi Pico
MOSI of LoRa module to GP19 of Raspberry Pi Pico
SCK of LoRa module to GP18 of Raspberry Pi Pico
NSS of LoRa module to GP17 of Raspberry Pi Pico
DIO0 of LoRa module to GP20 of Raspberry Pi Pico
Connect the Raspberry Pi Pico board to your computer using a USB cable.

Open a terminal program on your computer and connect to the Raspberry Pi Pico board with the following settings:

Baud rate: 115200
Data bits: 8
Stop bits: 1
Parity: None
Download and install the RadioHead library for Raspberry Pi Pico from GitHub. This library provides functions to control the LoRa transceiver module.

This program uses C++ using the RadioHead library to send and receive messages. 

This program initializes the LoRa transceiver module, sets the frequency to 915.0 MHz, and waits for input from the serial console. When it receives a message from the serial console, it sends the message over LoRa. When it receives a LoRa message, it prints the message to the serial console.

Compile and upload the program to the Raspberry Pi Pico board.

Open a serial console on your computer and send a message. The message should be received by the Raspberry Pi Pico board and printed to the serial console. The Raspberry Pi Pico board should also send an acknowledgment message back to the computer.

That's it! You have now sent and received messages using pi lora.